* src/quicc_smc_serial.h: Added two casts of (int) on calculations   involving CYGHWR_HAL_POWERPC_BOARD_SPEED.

+++ packages/devs/serial/powerpc/quicc/current/src/quicc_smc_serial.h
@@ -130,8 +130,8 @@ static cyg_int32 select_baud[] = {
     0,      // 230400
 };
 
-#define UART_BITRATE(n) ((((CYGHWR_HAL_POWERPC_BOARD_SPEED*1000000)/16)/n)-1)
-#define UART_SLOW_BITRATE(n) ((CYGHWR_HAL_POWERPC_BOARD_SPEED*1000000)/n))
+#define UART_BITRATE(n) ((((int)(CYGHWR_HAL_POWERPC_BOARD_SPEED*1000000)/16)/n)-1)
+#define UART_SLOW_BITRATE(n) ((int)(CYGHWR_HAL_POWERPC_BOARD_SPEED*1000000)/n))
 
 // Channel type select
 #define _SCC_CHAN 0
